\name{GTEX_LIVER_CRUDE}
\alias{GTEX_LIVER_CRUDE}
\alias{LIVER_EXPR}
\docType{data}
\title{
mRNA Expression Of Normal LIVER From GTEX
}
\description{
mRNA expression in LIVER
}
\usage{data(LIVER_EXPR)}
\format{
  A matrix
}

\source{
http://www.gtexportal.org
Genotype-Tissue Expression (GTEx)
}
\references{
Carithers et al.
Biopreservation and Biobanking. October 2015, 13(5): 311-319. doi:10.1089/bio.2015.0032.
PMID: 26484571. 
}
\examples{
data(LIVER_EXPR)
}
\keyword{datasets}
